Implemented job post creation logic: involves inter service communication with address controller(user service)

+++ jobspotter_backend/services/job-post/src/main/java/org/job_spotter/jobpost/model/JobTagEnum.java
@@ -2,36 +2,36 @@ package org.job_spotter.jobpost.model;
 
 public enum JobTagEnum {
 
-    GENERAL_HELP(1, "General Help"),
-    HANDYMAN_SERVICES(2, "Handyman Services"),
-    SKILLED_TRADES(3, "Skilled Trades"),
-    CLEANING_SERVICES(4, "Cleaning Services"),
-    DELIVERY_SERVICES(5, "Delivery Services"),
-    CAREGIVING(6, "Caregiving"),
-    PET_CARE(7, "Pet Care"),
-    TUTORING_MENTORING(8, "Tutoring/Mentoring"),
-    EVENT_STAFF(9, "Event Staff"),
-    ADMINISTRATIVE_SUPPORT(10, "Administrative Support"),
-    VIRTUAL_ASSISTANCE(11, "Virtual Assistance"),
-    FOOD_SERVICES(12, "Food Services"),
-    GARDENING_LANDSCAPING(13, "Gardening/Landscaping"),
-    COMMUNITY_OUTREACH(14, "Community Outreach"),
-    IT_SUPPORT(15, "IT Support"),
-    CREATIVE_SERVICES(16, "Creative Services"),
-    PERSONAL_SERVICES(17, "Personal Services"),
-    TUTORING_LANGUAGES(18, "Tutoring Languages"),
-    MUSIC_INSTRUCTION(19, "Music Instruction"),
-    OTHER(20, "Other");
+    GENERAL_HELP(1L, "General Help"),
+    HANDYMAN_SERVICES(2L, "Handyman Services"),
+    SKILLED_TRADES(3L, "Skilled Trades"),
+    CLEANING_SERVICES(4L, "Cleaning Services"),
+    DELIVERY_SERVICES(5L, "Delivery Services"),
+    CAREGIVING(6L, "Caregiving"),
+    PET_CARE(7L, "Pet Care"),
+    TUTORING_MENTORING(8L, "Tutoring/Mentoring"),
+    EVENT_STAFF(9L, "Event Staff"),
+    ADMINISTRATIVE_SUPPORT(10L, "Administrative Support"),
+    VIRTUAL_ASSISTANCE(11L, "Virtual Assistance"),
+    FOOD_SERVICES(12L, "Food Services"),
+    GARDENING_LANDSCAPING(13L, "Gardening/Landscaping"),
+    COMMUNITY_OUTREACH(14L, "Community Outreach"),
+    IT_SUPPORT(15L, "IT Support"),
+    CREATIVE_SERVICES(16L, "Creative Services"),
+    PERSONAL_SERVICES(17L, "Personal Services"),
+    TUTORING_LANGUAGES(18L, "Tutoring Languages"),
+    MUSIC_INSTRUCTION(19L, "Music Instruction"),
+    OTHER(20L, "Other");
 
-    private final int id;
+    private final Long id;
     private final String displayName;
 
-    JobTagEnum(int id, String displayName) {
+    JobTagEnum(Long id, String displayName) {
         this.id = id;
         this.displayName = displayName;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
